Address reviewer comments: Make BigDecimal tests solver-agnostic
